requirements.txt
asgiref==3.9.1
asttokens==3.0.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cssselect2==0.8.0
decorator==5.2.1
Django==4.2.23
django-cors-headers==3.14.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
Faker==37.4.2
fonttools==4.59.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
pyphen==0.17.2
requests==2.32.4
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tinyhtml5==2.0.0
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
